po2tmx allows us to convert PO files to TMX (Translation Memory eXchange) format. This format is designed to allow the sharing of TM data.
This implementation is beta alpha quality and has been sitting neglected on
my harddrive.  It is probably better that it move to a new home where someone
can take good care of it.

The TMX files it builds are probably wrong.  But at least if people can test it
then we can begin correcting errors.

setup.py now includes a new section for installing beta software.  Thus po2tmx
will not be installed by default but requires:

  installbeta = True

Enjoy!


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/trunk/translate@1187 54714841-351b-0410-a198-e36a94b762f5