Multiline comments in a dtd like this:
<!--
  Comment
 -->
<!ENTITY some.label "Blah">

End up losing the comment and all the following entities.  The culprit is
the space before the end comment marker.  If you remove that then it
extracts correctly.

This test uncovers the problem.  We're calling it line by line from the
parser in dtd.py and thus we are already in the string with we need to
extract.

This one is breaking our otherwise very good language packs.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/trunk/translate@2457 54714841-351b-0410-a198-e36a94b762f5