Merge branch/toolkit-C-po changes r6459:6622 into trunk.
These are only the PO related files.  All users of the PO classes
still need to be migrated.  The tools still use the Python PO
implementation but by using the environment variables USECPO=1 
you can force the tools to use the C PO version (USECPO=0 forces
the use of Python PO).  Using USECPO will report the implementation
being used, without USECPO operation continue as normal.

Common tests are in test_po while implementation specific test or
tests that differ slightly are in test_pypo and test_cpo.

The next steps will be to merge changes made to the tools the use 
po.py into trunk.  This will be done slowly to ensure that tests
and operation continue as normal.  On the branch almost all tests
are functioning correctly.  This will not be so on trunk until all
tests have been ported.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@6623 54714841-351b-0410-a198-e36a94b762f5