Removed the HORRIBLE hack whereby one could pass parameters to the stats code in order to get it not to throw an exception.
I ought to have my coding licence revoked. But at least until then, I
can try to restore my karma by restoring the code to a better state.

So anyway, if you call one of the stats functions on a non-existant or
illegal file, it's your responsibility to deal with the exceptions
that might pop out.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@7796 54714841-351b-0410-a198-e36a94b762f5