Add iCalendar conversion support: * This uses the vobject library to access .ics files. * We only process relevant text properties in VEVENT objects.   We could process other entries but why? * vobject should be >= 0.6.5
On a positive note, these where implemented based on ini.py 
and the limited amount that needed to be implemented for 
ical.py and po2ical/ical2po highlighted that we are close
to making this generic enough that by simply adding a format
we get a converter at not extra cost.

Happy localising of Calendars!


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@8461 54714841-351b-0410-a198-e36a94b762f5