Fixed possible stats database inconsistency problems. Also modified statsdb never to catch any exceptions. The user code must handle exceptions.
In the previous code, if an error occurred before a database commit
was issued, then a next database commit would pull in possibly
inconsistent changes from the previous failed call.

To ensure consistency, the database MUST be rolled back if ANY
exception, whatsoever is raised in the database code. Why? Because
it's impossible to know whether the database state is consistent
at the point when an exception is thrown.

The transaction decorator will ensure a database commit if
a decorated function executes without problems. Otherwise (if
an exception occured), it will roll back the database and
reraise the exception.

Also note that pocount now handles exceptions from statsdb.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@8791 54714841-351b-0410-a198-e36a94b762f5