In the process of replacing placeables support for the toolkit.
The toolkit's placeable support is based on XLIFF's placeables.
There are structures defined for all of the XLIFF placeables.

The support is currently limited. We don't properly support
marked content yet. We also don't enforce constraints such
as that <sub> tags cannot be children of <g> tags. This will
all follow in the future.

This patch includes placeable support for the LISA store types.
The accessors get_rich_source and set_rich_source are used to
set the source of a unit with placeables. Likewise the
accessors get_rich_target and set_rich_target are used to
access the target of a unit.

See the unit tests for how these should be used.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@8834 54714841-351b-0410-a198-e36a94b762f5