First, having
  if target == self.target:

in gettarget slows things down enormously, since self.target is not that cheap.

When I removed this, I uncovered a bug in a test. That's fixed now.

This also pointed to a bit of incompleteness in setsource, which is now fixed.

Finally, we don't have to call settarget in the constructor, since the base
class does it.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/translate@9238 54714841-351b-0410-a198-e36a94b762f5